requirements.txt
# BAC Tracker — web app & CLI
flask>=3.0.0
PyYAML>=6.0
matplotlib>=3.7.0
gunicorn>=21.0.0
pytest>=7.0.0
